move validation logic to validation file
